More exports from index.tsx. Also fixes up build errors.
