+++ src/ui/common/src/components/primitives/Tabs.styles.ts
@@ -1,8 +1,9 @@
 import { styled } from '@mui/material/styles';
 import Tab, { tabClasses } from '@mui/material/Tab';
 import Tabs, { tabsClasses } from '@mui/material/Tabs';
+import { theme } from "../../styles/theme/theme";
 
-const AqueductTabs = styled(Tabs)(({ theme }) => {
+const AqueductTabs = styled(Tabs)(() => {
   return {
     [`& .${tabsClasses.indicator}`]: {
       backgroundColor: theme.palette.blue[800],
@@ -10,7 +11,7 @@ const AqueductTabs = styled(Tabs)(({ theme }) => {
   };
 });
 
-const AqueductTab = styled(Tab)(({ theme }) => {
+const AqueductTab = styled(Tab)(() => {
   return {
     [`&.${tabClasses.selected}`]: {
       color: theme.palette.blue[800],
