+++ src/ui/common/src/styles/theme/theme.tsx
@@ -1,74 +1,74 @@
 export const theme = {
-  palette: {
-    black: '#000000',
-    white: '#ffffff',
-    darkGray: '#333333',
-    gray: {
-      900: '#101840',
-      800: '#474d66',
-      700: '#696f8c',
-      600: '#8f95b2',
-      500: '#c1c4d6',
-      400: '#d8dae5',
-      300: '#E6E8F0',
-      200: '#edeff5',
-      100: '#F4F5F9',
-      90: '#F4F6FA',
-      75: '#F9FAFC',
-      50: '#F2F2F2',
+    palette: {
+        black: '#000000',
+        white: '#ffffff',
+        darkGray: '#333333',
+        gray: {
+            900: '#101840',
+            800: '#474d66',
+            700: '#696f8c',
+            600: '#8f95b2',
+            500: '#c1c4d6',
+            400: '#d8dae5',
+            300: '#E6E8F0',
+            200: '#edeff5',
+            100: '#F4F5F9',
+            90: '#F4F6FA',
+            75: '#F9FAFC',
+            50: '#F2F2F2',
+        },
+        blue: {
+            900: '#002F5E',
+            800: '#004080',
+            700: '#0059B3',
+            500: '#0073E6',
+            400: '#0080FF',
+            300: '#4DA6FF',
+            100: '#CCE6FF',
+            50: '#E6F2FF',
+        },
+        red: {
+            800: '#611F1F',
+            700: '#7D2828',
+            600: '#A73636',
+            500: '#D14343',
+            300: '#EE9191',
+            100: '#F9DADA',
+            25: '#FDF4F4',
+        },
+        green: {
+            900: '#10261E',
+            800: '#214C3C',
+            700: '#317159',
+            600: '#429777',
+            500: '#52BD95',
+            400: '#75CAAA',
+            300: '#97D7BF',
+            200: '#BAE5D5',
+            100: '#DCF2EA',
+            25: '#F5FBF8',
+        },
+        orange: {
+            700: '#996A13',
+            600: '#FFA600',
+            500: '#FFB020',
+            100: '#F8E3DA',
+            25: '#FFFAF2',
+        },
+        purple: {
+            600: '#6E62B6',
+            100: '#E7E4F9',
+        },
+        teal: {
+            800: '#0F5156',
+            100: '#D3F5F7',
+        },
+        yellow: {
+            800: '#66460D',
+            100: '#FFEFD2',
+        },
+        TableSuccessBackground: 'rgba(76,175,80,0.1)',
+        TableErrorBackground: 'rgba(224,67,54,0.1)',
+        TableWarningBackground: 'rgba(237,18,2,0.1)',
     },
-    blue: {
-      900: '#002F5E',
-      800: '#004080',
-      700: '#0059B3',
-      500: '#0073E6',
-      400: '#0080FF',
-      300: '#4DA6FF',
-      100: '#CCE6FF',
-      50: '#E6F2FF',
-    },
-    red: {
-      800: '#611F1F',
-      700: '#7D2828',
-      600: '#A73636',
-      500: '#D14343',
-      300: '#EE9191',
-      100: '#F9DADA',
-      25: '#FDF4F4',
-    },
-    green: {
-      900: '#10261E',
-      800: '#214C3C',
-      700: '#317159',
-      600: '#429777',
-      500: '#52BD95',
-      400: '#75CAAA',
-      300: '#97D7BF',
-      200: '#BAE5D5',
-      100: '#DCF2EA',
-      25: '#F5FBF8',
-    },
-    orange: {
-      700: '#996A13',
-      600: '#FFA600',
-      500: '#FFB020',
-      100: '#F8E3DA',
-      25: '#FFFAF2',
-    },
-    purple: {
-      600: '#6E62B6',
-      100: '#E7E4F9',
-    },
-    teal: {
-      800: '#0F5156',
-      100: '#D3F5F7',
-    },
-    yellow: {
-      800: '#66460D',
-      100: '#FFEFD2',
-    },
-    TableSuccessBackground: 'rgba(76,175,80,0.1)',
-    TableErrorBackground: 'rgba(224,67,54,0.1)',
-    TableWarningBackground: 'rgba(237,18,2,0.1)',
-  },
 };
