+++ src/ui/common/tsdx.config.js
@@ -1,0 +1,12 @@
+const postcss = require('rollup-plugin-postcss');
+
+module.exports = {
+    rollup(config, options) {
+        config.plugins.push(
+            postcss({
+                modules: true,
+            })
+        );
+        return config;
+    },
+};
